+++ HISTORY.rst
@@ -7,3 +7,5 @@ History
 
 * First release
 * Add an API client to wrap the Data for SEO RestClient
+* Add ability to get credentials from environment variables
+** DFS_API_USERNAME and DFS_API_PASSWORD